+++ Assets/AppsFlyer/AppsFlyer.cs
@@ -7,7 +7,7 @@ namespace AppsFlyerSDK
     public class AppsFlyer : MonoBehaviour
     {
 
-        public static readonly string kAppsFlyerPluginVersion = "6.9.4";
+        public static readonly string kAppsFlyerPluginVersion = "6.10.10";
         public static string CallBackObjectName = null;
         private static EventHandler onRequestResponse;
         private static EventHandler onInAppResponse;
@@ -295,9 +295,16 @@ namespace AppsFlyerSDK
             if (instance != null)
             {
                 instance.setCurrencyCode(currencyCode);
-            }
-
-
+            } else {
+#if UNITY_IOS || UNITY_STANDALONE_OSX
+                instance = new AppsFlyeriOS();
+                instance.setCurrencyCode(currencyCode);
+#elif UNITY_ANDROID
+                instance = new AppsFlyerAndroid();
+                instance.setCurrencyCode(currencyCode);
+#else
+#endif
+        }
         }
 
         /// <summary>
@@ -378,9 +385,16 @@ namespace AppsFlyerSDK
             if (instance != null)
             {
                 instance.setHost(hostPrefixName, hostName);
-            }
-
-
+            } else {
+#if UNITY_IOS || UNITY_STANDALONE_OSX
+                instance = new AppsFlyeriOS();
+                instance.setHost(hostPrefixName, hostName);
+#elif UNITY_ANDROID
+                instance = new AppsFlyerAndroid();
+                instance.setHost(hostPrefixName, hostName);
+#else
+#endif
+        }
         }
 
         /// <summary>
@@ -582,8 +596,8 @@ namespace AppsFlyerSDK
         {
             if (instance != null && instance is IAppsFlyerIOSBridge)
             {
-                IAppsFlyerIOSBridge appsFlyerAndroidInstance = (IAppsFlyerIOSBridge)instance;
-                appsFlyerAndroidInstance.setDisableCollectAppleAdSupport(disable);
+                IAppsFlyerIOSBridge appsFlyeriOSInstance = (IAppsFlyerIOSBridge)instance;
+                appsFlyeriOSInstance.setDisableCollectAppleAdSupport(disable);
             }
         }
 
@@ -591,8 +605,8 @@ namespace AppsFlyerSDK
         {
             if (instance != null && instance is IAppsFlyerIOSBridge)
             {
-                IAppsFlyerIOSBridge appsFlyerAndroidInstance = (IAppsFlyerIOSBridge)instance;
-                appsFlyerAndroidInstance.setShouldCollectDeviceName(shouldCollectDeviceName);
+                IAppsFlyerIOSBridge appsFlyeriOSInstance = (IAppsFlyerIOSBridge)instance;
+                appsFlyeriOSInstance.setShouldCollectDeviceName(shouldCollectDeviceName);
             }
         }
 
@@ -634,8 +648,8 @@ namespace AppsFlyerSDK
         {
             if (instance != null && instance is IAppsFlyerIOSBridge)
             {
-                IAppsFlyerIOSBridge appsFlyerAndroidInstance = (IAppsFlyerIOSBridge)instance;
-                appsFlyerAndroidInstance.setDisableCollectIAd(disableCollectIAd);
+                IAppsFlyerIOSBridge appsFlyeriOSInstance = (IAppsFlyerIOSBridge)instance;
+                appsFlyeriOSInstance.setDisableCollectIAd(disableCollectIAd);
             }
         }
 
@@ -653,8 +667,8 @@ namespace AppsFlyerSDK
         {
             if (instance != null && instance is IAppsFlyerIOSBridge)
             {
-                IAppsFlyerIOSBridge appsFlyerAndroidInstance = (IAppsFlyerIOSBridge)instance;
-                appsFlyerAndroidInstance.setUseReceiptValidationSandbox(useReceiptValidationSandbox);
+                IAppsFlyerIOSBridge appsFlyeriOSInstance = (IAppsFlyerIOSBridge)instance;
+                appsFlyeriOSInstance.setUseReceiptValidationSandbox(useReceiptValidationSandbox);
             }
         }
 
@@ -679,8 +693,8 @@ namespace AppsFlyerSDK
         {
             if (instance != null && instance is IAppsFlyerIOSBridge)
             {
-                IAppsFlyerIOSBridge appsFlyerAndroidInstance = (IAppsFlyerIOSBridge)instance;
-                appsFlyerAndroidInstance.setUseUninstallSandbox(useUninstallSandbox);
+                IAppsFlyerIOSBridge appsFlyeriOSInstance = (IAppsFlyerIOSBridge)instance;
+                appsFlyeriOSInstance.setUseUninstallSandbox(useUninstallSandbox);
             }
         }
 
@@ -707,8 +721,8 @@ namespace AppsFlyerSDK
         {
             if (instance != null && instance is IAppsFlyerIOSBridge)
             {
-                IAppsFlyerIOSBridge appsFlyerAndroidInstance = (IAppsFlyerIOSBridge)instance;
-                appsFlyerAndroidInstance.validateAndSendInAppPurchase(productIdentifier, price, currency, tranactionId, additionalParameters, gameObject);
+                IAppsFlyerIOSBridge appsFlyeriOSInstance = (IAppsFlyerIOSBridge)instance;
+                appsFlyeriOSInstance.validateAndSendInAppPurchase(productIdentifier, price, currency, tranactionId, additionalParameters, gameObject);
             }
         }
 
@@ -725,8 +739,8 @@ namespace AppsFlyerSDK
         { 
             if (instance != null && instance is IAppsFlyerIOSBridge)
             {
-                IAppsFlyerIOSBridge appsFlyerAndroidInstance = (IAppsFlyerIOSBridge)instance;
-                appsFlyerAndroidInstance.handleOpenUrl(url, sourceApplication, annotation);
+                IAppsFlyerIOSBridge appsFlyeriOSInstance = (IAppsFlyerIOSBridge)instance;
+                appsFlyeriOSInstance.handleOpenUrl(url, sourceApplication, annotation);
             }
         }
 
@@ -734,8 +748,8 @@ namespace AppsFlyerSDK
         {
             if (instance != null && instance is IAppsFlyerIOSBridge)
             {
-                IAppsFlyerIOSBridge appsFlyerAndroidInstance = (IAppsFlyerIOSBridge)instance;
-                appsFlyerAndroidInstance.registerUninstall(deviceToken);
+                IAppsFlyerIOSBridge appsFlyeriOSInstance = (IAppsFlyerIOSBridge)instance;
+                appsFlyeriOSInstance.registerUninstall(deviceToken);
             }
         }
 
@@ -743,8 +757,8 @@ namespace AppsFlyerSDK
         {
             if (instance != null && instance is IAppsFlyerIOSBridge)
             {
-                IAppsFlyerIOSBridge appsFlyerAndroidInstance = (IAppsFlyerIOSBridge)instance;
-                appsFlyerAndroidInstance.waitForATTUserAuthorizationWithTimeoutInterval(timeoutInterval);
+                IAppsFlyerIOSBridge appsFlyeriOSInstance = (IAppsFlyerIOSBridge)instance;
+                appsFlyeriOSInstance.waitForATTUserAuthorizationWithTimeoutInterval(timeoutInterval);
             }
         }
 
@@ -752,8 +766,8 @@ namespace AppsFlyerSDK
         {
             if (instance != null && instance is IAppsFlyerIOSBridge)
             {
-                IAppsFlyerIOSBridge appsFlyerAndroidInstance = (IAppsFlyerIOSBridge)instance;
-                appsFlyerAndroidInstance.setCurrentDeviceLanguage(language);
+                IAppsFlyerIOSBridge appsFlyeriOSInstance = (IAppsFlyerIOSBridge)instance;
+                appsFlyeriOSInstance.setCurrentDeviceLanguage(language);
             }
         }
 
@@ -776,9 +790,16 @@ namespace AppsFlyerSDK
         {
             if (instance != null && instance is IAppsFlyerIOSBridge)
             {
-                IAppsFlyerIOSBridge appsFlyerAndroidInstance = (IAppsFlyerIOSBridge)instance;
-                appsFlyerAndroidInstance.disableSKAdNetwork(isDisabled);
-            }
+                IAppsFlyerIOSBridge appsFlyeriOSInstance = (IAppsFlyerIOSBridge)instance;
+                appsFlyeriOSInstance.disableSKAdNetwork(isDisabled);
+            } else {
+#if UNITY_IOS || UNITY_STANDALONE_OSX
+                instance = new AppsFlyeriOS();
+                IAppsFlyerIOSBridge appsFlyeriOSInstance = (IAppsFlyerIOSBridge)instance;
+                appsFlyeriOSInstance.disableSKAdNetwork(isDisabled);
+#else
+#endif
+        }
         }
 
         public static void setCollectOaid(bool isCollect)