+++ Assets/AppsFlyer/Windows/AppsFlyerWindows.cs
@@ -77,8 +77,12 @@ namespace AppsFlyerSDK
             Log("[Callback]: {0}", code.ToString());
 
             AppsFlyerRequestEventArgs eventArgs = new AppsFlyerRequestEventArgs((int)code, code.ToString());
-            var method = _gameObject.GetType().GetMethod("AppsFlyerOnRequestResponse");
-            method.Invoke(_gameObject, new object[] { AppsFlyerTracker.GetAppsFlyerTracker(), eventArgs });
+            if (_gameObject != null) {
+                var method = _gameObject.GetType().GetMethod("AppsFlyerOnRequestResponse");
+                if (method != null) {
+                    method.Invoke(_gameObject, new object[] { AppsFlyerTracker.GetAppsFlyerTracker(), eventArgs });
+                }
+            }
         }
 #endif